+++ packages/react-sdk/src/lib/index.ts
@@ -27,4 +27,4 @@ export { FontsLoadedContextProvider, useFontsLoaded } from './useFontsLoaded';
 export { useLatestValue } from './useLatestValue';
 export { useMeasure } from './useMeasure';
 export { getUserColor } from './userColor';
-export { useZoomControls } from './zoom';
+export { useZoomControls } from './useZoomControls';
